+++ surfsense_backend/app/tasks/stream_connector_search_results.py
@@ -3,7 +3,7 @@ from uuid import UUID
 
 from app.agents.researcher.graph import graph as researcher_graph
 from app.agents.researcher.state import State
-from app.utils.streaming_service import StreamingService
+from app.services.streaming_service import StreamingService
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.agents.researcher.configuration import SearchMode
